change int to float

+++ Assets/Scripts/Food.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class Food : MonoBehaviour
 {
-	public int FoodValue = 0;
+	public float FoodValue = 0;
 	public AudioClip EatingSound;
 }
